Crate wdk

Idiomatic Rust wrappers for the Windows Driver Kit (WDK) APIs. This crate is built on top of the raw FFI bindings provided by [wdk-sys], and provides a safe, idiomatic rust interface to the WDK.

Re-exports

pub use print::_print;

Macros

paged_code

print

print to kernel debugger via wdk_sys::ntddk::DbgPrint

println

print with newline to debugger via wdk_sys::ntddk::DbgPrint

Functions

dbg_break

Trigger a breakpoint in debugger via architecture-specific inline assembly.

nt_success

Evaluates to TRUE if the return value specified by nt_status is a success type (0 − 0x3FFFFFFF) or an informational type (0x40000000 − 0x7FFFFFFF). This function is taken from ntdef.h in the WDK.